init commit

+++ src/BackupManager.php
@@ -336,49 +336,39 @@ class BackupManager
                 Storage::disk($this->disk)->delete($backupPath);
             }
 
-            // Use a try-catch block for better error handling
-            try {
-                DB::table('verifybackup')->updateOrInsert(['id' => 1], ['verify_status' => 'backup']);
+            DB::table('verifybackup')->updateOrInsert(
+                ['id' => 1],
+                ['verify_status' => 'backup']
+            );
 
-                $connection = config('database.connections.mysql');
-                $connectionOptions = "--user={$connection['username']} --password=\"{$connection['password']}\" --host={$connection['host']} {$connection['database']} ";
-                $options = "--single-transaction --max-allowed-packet=4096 --quick";
+            $databaseConfig = config('database.connections.mysql');
+            $dbUsername = $databaseConfig['username'];
+            $dbPassword = $databaseConfig['password'];
+            $dbHost = $databaseConfig['host'];
+            $dbName = $databaseConfig['database'];
 
-                $itemsToBackup = config('lara-backup-manager.backups.database.tables');
-                if ($itemsToBackup) {
-                    $itemsToBackup[] = 'verifybackup';
-                    $tableOptions = implode(' ', $itemsToBackup);
-                } else {
-                    $tableOptions = '';
-                }
+            $command = "mysqldump --user={$dbUsername} --password={$dbPassword} --host={$dbHost} {$dbName} | gzip > $this->dbBackupName";
 
-                $command = "cd " . str_replace('\\', '/', base_path()) . " && {$this->mysqldump} {$options} {$connectionOptions} {$tableOptions} | gzip > {$this->dbBackupName}";
-                // Add error handling for shell execution
-                $output = shell_exec($command . ' 2>&1');
-                if ($output !== null) {
-                    throw new Exception("Error while executing the backup command: $output");
-                }
+            exec($command);
 
-                // Move the backup file to the storage disk
-                $localFilePath = base_path($this->dbBackupName);
+            if (file_exists(base_path($this->dbBackupName))) {
                 $storageLocal = Storage::createLocalDriver(['root' => base_path()]);
-              return  $file = $storageLocal->get($this->dbBackupName);
-                Storage::disk($this->disk)->put($backupPath, $file);
+                $file = $storageLocal->get($this->dbBackupName);
 
-                // Delete the local backup file
+                Storage::disk($this->disk)->put($this->backupPath . $this->dbBackupName, $file);
+
+                // delete local file
                 $storageLocal->delete($this->dbBackupName);
 
-                // If this is a bypass operation, clean up old backups and return the status
-                if ($bypass === true) {
-                    $this->deleteOldBackups("db");
-                    return $this->getBackupStatus('db');
-                }
-            } catch (Exception $e) {
-                // Log and handle the exception
-                Log::error("Database backup failed: " . $e->getMessage());
-                // Optionally, re-throw the exception if you want to handle it at a higher level
-                // throw $e;
             }
+
+            if ($bypass===true) {
+                $this->deleteOldBackups("db");
+
+                return $this->getBackupStatus('db');
+
+            }
+
         }
     }
     protected function restoreFiles($file): void
